Deploy (Version 1.3): First Play Store version
Added full support to firebase analytics. Added the the app ads. Solving some problems in the UI and improved the performance. Added the loading arrow feedback for better UX. Created the banner image and the app icon image to publish the app on play store. Added the rate my app and privacy policy menu options.